+++ config/application.rb
@@ -21,5 +21,7 @@ module ZazoFriendfinder
 
     config.autoload_paths << Rails.root.join('lib')
     config.autoload_paths += %W(#{config.root}/app/queries)
+
+    config.active_record.raise_in_transactional_callbacks = true
   end
 end